requirements.txt
django==5.2.8
python-dotenv==1.2.1
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
requests==2.32.5
gunicorn==23.0.0
pillow==12.0.0
django-q2==1.9.0
PyJWT==2.10.1
stripe==14.1.0
cryptography==46.0.3
django-ratelimit==4.1.0
celery==5.6.1
redis==7.1.0
firebase_admin==7.1.0
python-dateutil==2.9.0.post0
twilio==9.10.2
psycopg2-binary==2.9.11
pytest==9.0.2
pytest-django==4.12.0
factory_boy==3.3.3
resend==2.29.0
reportlab==4.4.10
pypdf==5.1.0